pyproject.toml
[project]
name = "CuBATS"
version = "0.1.0"
description = "Computer Vision Based Antigen Target Selector"
authors = [
    { name = "Moritz Dinser", email = "moritz.dinser@hnu.de" },
    { name = "Daniel Hieber", email = "daniel.hieber@hnu.de" },
]
dependencies = [
    "openslide-python>=1.3.1",
    "valis-wsi>=1.0.4",
    "scikit-image==0.19.3",
    "onnx>=1.16.0",
    "onnx2torch==1.5.13",
    "onnxruntime-gpu>=1.10.0",
    "sphinx>=7.3.7",
    "sphinx-rtd-theme>=2.0.0",
    "torchvision>=0.20.0",
    "torchstain>=1.3.0",
    "cupy-cuda12x>=13.3.0",
]
requires-python = ">=3.10,<=3.12"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
testing = [
    "pytest>=8.2.0",
    "coverage>=7.5.0",
    "flake8>=7.0.0",
    "pylint>=3.2.7",
    "graphviz>=0.20.3",
    "imagecodecs>=2024.6.1",
]

[tool.pdm.scripts]
test = "pdm run -v pytest tests"
lint = "pdm run -v flake8 --ignore=E124,E127,E128,E203,E701,E731,W503 --max-line-length 120 src tests"
sort = "pdm run -v isort --check-only src tests"
apply_sort = "pdm run -v isort src tests"
cov_report = { shell = "pdm run -v coverage run -m pytest tests && pdm run coverage xml && pdm run coverage report" }
check_pre_commit = "pdm run pre-commit run --all-files"
# run prior to tests
pre_test = "pdm install"
build-docs = { shell = """
cd docs && make clean && make html
if [ "$(uname)" == "Darwin" ]; then
    open build/html/index.html
elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
    xdg-open build/html/index.html
elif [ "$(expr substr %(uname -s) 1 10" == "MINGW32_NT" ] || [ "$(expr substr %(uname -s) 1 10" == "MINGW64_NT" ]; then
    start /build/html/index.html
fi
"""}

[tool.isort]
src_paths = ["src", "tests"]
known_first_party = "cubats"
import_heading_stdlib = "Standard Library"
import_heading_thirdparty = "Third Party"
import_heading_firstparty = "CuBATS"

[dependency-groups]
dev = [
    "pre-commit>=3.7.0",
    "notebook>=7.1.3",
    "isort>=5.13.2",
    "wheel>=0.45.1",
]
testing = [
    "pytest>=8.2.0",
    "coverage>=7.5.0",
    "flake8>=7.0.0",
    "pylint>=3.2.7",
    "graphviz>=0.20.3",
    "imagecodecs>=2024.6.1",
]
